Implement generic Media streaming parser for  Audio & Transcription Data
1> Extend Audio & Transcription websocket models from common MediaBase
2> MediaParser is now a generic parser for all media streaming for Audio, Transcription and any other future addition
3> Fix namespaces for Dialog models & evetns
4> Re-generate model & create api-view
5> Refactor Audio & Transcription model test cases
