+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/MediaStreaming/MediaParser.cs
@@ -1,0 +1,86 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Text;
+using System.Text.Json;
+
+namespace Azure.Communication.CallAutomation
+{
+    /// <summary>
+    /// Parser for different packages, such as Audio data or Transcription, received as
+    /// part of Media Streaming
+    /// </summary>
+    public static class MediaParser
+    {
+        /// <summary>
+        /// Parsing a MediaStreaming package from BinaryData.
+        /// </summary>
+        /// <param name="json"></param>
+        /// <returns></returns>
+        public static MediaBase Parse(BinaryData json)
+        {
+            return Parse(json.ToString());
+        }
+
+        /// <summary>
+        /// Parsing a MediaStreaming package from a byte array.
+        /// </summary>
+        /// <param name="receivedBytes">a UTF8 byte array.</param>
+        /// <returns></returns>
+        public static MediaBase Parse(byte[] receivedBytes)
+        {
+            return Parse(Encoding.UTF8.GetString(receivedBytes));
+        }
+
+        /// <summary>
+        /// Parse the incoming package.
+        /// </summary>
+        /// <param name="stringJson"></param>
+        /// <returns></returns>
+        /// <exception cref="NotImplementedException"></exception>
+        public static MediaBase Parse(string stringJson)
+        {
+            JsonElement package = JsonDocument.Parse(stringJson).RootElement;
+
+            string kind = package.GetProperty("kind").ToString();
+
+            switch (kind)
+            {
+                #region Audio
+                case "AudioMetadata":
+                    return JsonSerializer.Deserialize<AudioMetadata>(package.GetProperty("audioMetadata").ToString());
+
+                case "AudioData":
+                    AudioDataInternal audioInternal = JsonSerializer.Deserialize<AudioDataInternal>(package.GetProperty("audioData").ToString());
+                    return new AudioData(
+                        audioInternal.Data, audioInternal.Timestamp, audioInternal.ParticipantRawId, audioInternal.Silent);
+
+                #endregion
+
+                #region Transcription
+                case "TranscriptionMetadata":
+                    return JsonSerializer.Deserialize<TranscriptionMetadata>(package.GetProperty("transcriptionMetadata").ToString());
+
+                case "TranscriptionData":
+                    TranscriptionDataInternal transcriptionDataInternal = JsonSerializer.Deserialize<TranscriptionDataInternal>(
+                   package.GetProperty("transcriptionData").ToString()
+                   );
+                    return new TranscriptionData(
+                        transcriptionDataInternal.Text,
+                        transcriptionDataInternal.Format,
+                        transcriptionDataInternal.Confidence,
+                        transcriptionDataInternal.Offset,
+                        transcriptionDataInternal.Words,
+                        transcriptionDataInternal.ParticipantRawID,
+                        transcriptionDataInternal.ResultStatus
+                        );
+
+                    #endregion
+
+                default:
+                    throw new NotSupportedException(stringJson);
+            }
+        }
+    }
+}
