+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/MediaStreaming/Transcription/ResultStatus.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Azure.Communication.CallAutomation.Models.Transcription
+namespace Azure.Communication.CallAutomation
 {
     /// <summary>
     /// The status of the result of transcription
