--- sdk/communication/Azure.Communication.CallAutomation/src/Models/Transcription/TextFormat.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Azure.Communication.CallAutomation.Models.Transcription
-{
-    /// <summary>
-    /// The format of transcription text
-    /// </summary>
-    public enum TextFormat
-    {
-        /// <summary>
-        ///  Formatted recognize text with punctuations
-        /// </summary>
-        Display
-    }
-}
